pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Fix_agent"
version = "1.0.0"
description = "AI Coding Assistant - 代码缺陷修复和分析工具"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MPL-2.0"}
authors = [
    {name = "3uyuan1ee", email = "1481059602@qq.com"}
]
maintainers = [
    {name = "3uyuan1ee", email = "1481059602@qq.com"}
]
keywords = [
    "ai", "coding", "assistant", "cli", "code-analysis",
    "bug-detection", "deepagents", "code-review", "automated-testing",
    "code-quality", "static-analysis", "python", "javascript", "java"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Operating System :: OS Independent",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Debuggers",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Environment :: Console",
]

dependencies = [
    # CLI和UI
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.52",
    "requests>=2.28.0",

    # AI和ML依赖
    "langchain-openai>=0.1.0",
    "langchain-core>=0.1.0",
    "deepagents==0.2.4",
    "python-dotenv>=1.0.0",
    "langgraph>=0.1.0",

    # 联网搜索
    "tavily-python>=0.3.0",

    # 性能监控
    "psutil>=5.9.0",

    # 核心Python代码分析工具 - 确保开箱即用
    "pylint>=2.17.0",         # 全面的Python代码分析
    "flake8>=6.0.0",           # Python代码风格和基础检查
    "mypy>=1.0.0",             # Python静态类型检查
    "bandit>=1.7.0",           # Python安全漏洞检查
    "black>=23.0.0",           # Python代码格式化
    "astroid>=2.15.0",         # pylint依赖
    "isort>=5.12.0",           # 导入排序工具
    "pydocstyle>=6.0.0",      # 文档字符串风格检查

    # 基础解析和配置支持
    "lxml>=4.9.0",             # XML解析（用于工具输出解析）
    "pyyaml>=6.0.0",           # YAML配置文件解析
    "toml>=0.10.0",            # TOML配置文件解析

    # 工具生成的测试依赖
    "pytest>=7.0.0",           # 测试生成器需要
    "typing-extensions>=4.0.0", # 类型扩展支持
]

[project.optional-dependencies]
# 测试工具 - 仅用于开发和测试环境
test = [
    "pytest>=7.0.0",           # 测试框架
    "pytest-cov>=4.0.0",       # 测试覆盖率
    "pytest-mock>=3.10.0",     # Mock支持
    "pytest-asyncio>=0.21.0",  # 异步测试支持
    "pytest-xdist>=3.3.0",     # 并行测试
]

# 开发工具 - 仅用于代码开发和质量控制
dev = [
    "Fix_agent[test]",          # 包含测试依赖
    "black>=23.0.0",            # 代码格式化
    "flake8>=6.0.0",            # 代码风格检查
    "mypy>=1.0.0",              # 静态类型检查
    "pre-commit>=3.0.0",        # Git钩子
]

# 发布工具 - 仅用于打包和发布
release = [
    "Fix_agent[dev]",           # 包含开发依赖
    "twine>=4.0.0",             # PyPI上传工具
    "wheel>=0.41.0",            # 打包工具
    "build>=0.10.0",            # 构建工具
]

# 文档工具 - 仅用于文档生成
docs = [
    "sphinx>=5.0.0",                    # 文档生成
    "sphinx-rtd-theme>=1.2.0",          # RTD主题
    "myst-parser>=0.18.0",              # Markdown支持
    "sphinx-autodoc-typehints>=1.22.0", # 类型提示支持
]

# 完整开发环境 - 包含所有开发工具
full = [
    "Fix_agent[test,dev,release,docs]",
]

# Windows构建工具
windows = [
    "Fix_agent[release]",
    "pywin32>=226",  # Windows API支持
    "wmi>=1.5.1",    # Windows管理接口
]

# Web应用依赖
web = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "psutil>=5.9.0",
]

[project.urls]
Homepage = "https://github.com/3uyuan1ee/Fix_agent"
Documentation = "https://github.com/3uyuan1ee/Fix_agent#readme"
Repository = "https://github.com/3uyuan1ee/Fix_agent"
"Bug Tracker" = "https://github.com/3uyuan1ee/Fix_agent/issues"
"Changelog" = "https://github.com/3uyuan1ee/Fix_agent/blob/main/CHANGELOG.md"

[project.scripts]
fix-agent = "src.main:cli_main"
fixagent = "src.main:cli_main"
fixagent-web = "src.web_app_main:cli_main"
fixagent-web-dev = "src.web_app_main:web_main"
fixagent-web-stop = "src.web_app_stop:cli_main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "web_app*"]

[tool.setuptools.package-data]
src = ["*.md", "*.txt", "*.yml", "*.yaml", "*.json", "*.toml"]
web_app = ["**/*.html", "**/*.css", "**/*.js", "**/*.json", "**/*.png", "**/*.jpg", "**/*.ico", "*.md", "*.txt"]
"*" = [".env.template","src/**/*.md", "src/**/*.txt"]


[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]